Added mixer for differential steering and RC PWM deadband
The mixer allows the use of a pistol-grip or single-stick RC transmitter for controlling skid-steer or tank-style robots with the ODrive.

The steering mixer uses Fibre and remote properties for configuration just like the PWM pins do.  It's designed for minimal impact on the codebase.

The PWM deadband addition places a 10-microsecond-wide band in the middle of the PWM pulse time range in which the center of the channel's min and max setting is output.  I'm not sure at the moment if this feature should have a toggle to disable.