Fix hang on compile on Windows
Windows 10 includes since version 1903 a python stub which is supposed
to open the MS Store if python is invoked but not installed. However
when invoked from tup, the stub simply hangs without output.

With this change we first check the python version (something that the
Windows stub does not interer with) before we invoke it.