reduce likelihood of `IndexError` during DFU
In some cases `DfuDevice.control_msg` returns zero bytes. This might be
because the device didn't have enough time to process the previous
command. This commit adds a single retry after a delay. If the retry
fails too, an error message is printed that helps the user to recover
the ODrive.

Since this issue occurs rarely it's unclear if this commit really solves
the issue.